requirements.txt
mlflow
streamli
pandas
plotly
scikit-learn
pyyaml
numpy
cryptography
cffi
pycparser
